sending an array on events and user endpoionts instead of an object, adapting front to this change
