+++ src/api/admins.py
@@ -1,9 +1,9 @@
 from flask import Flask, request, jsonify, url_for, Blueprint
 
-admin = Blueprint('admin', __name__,)
+admins = Blueprint('admins', __name__,)
 
 
 
-@admin.route('/hello', methods=['GET'])
+@admins.route('/hello', methods=['GET'])
 def hello():
     return jsonify({"message": "Bien"}), 200
